Refactor tests for JSON_EXTRACT, JSON_INSERT, JSON_VALUE, and JSON_CONTAINS_PATH so that we can test every JSONWrapper implementation, and run these tests from Dolt.
