+++ sql/expression/function/json/jsontests/json_function_tests.go
@@ -15,7 +15,6 @@
 package jsontests
 
 import (
-	"fmt"
 	"strconv"
 	"strings"
 	"testing"
@@ -24,11 +23,10 @@ import (
 
 	"github.com/dolthub/go-mysql-server/sql"
 	"github.com/dolthub/go-mysql-server/sql/expression"
-	"github.com/dolthub/go-mysql-server/sql/expression/function/json"
 	"github.com/dolthub/go-mysql-server/sql/types"
 )
 
-type prepareJsonValue = func(*testing.T, string) interface{}
+type prepareJsonValue = func(*testing.T, interface{}) interface{}
 
 type jsonFormatTest struct {
 	name        string
@@ -38,13 +36,15 @@ type jsonFormatTest struct {
 var jsonFormatTests = []jsonFormatTest{
 	{
 		name: "string",
-		prepareFunc: func(t *testing.T, js string) interface{} {
-			return js
+		prepareFunc: func(t *testing.T, js interface{}) interface{} {
+			jsonString, _, err := types.Text.Convert(js)
+			require.NoError(t, err)
+			return jsonString
 		},
 	},
 	{
 		name: "JsonDocument",
-		prepareFunc: func(t *testing.T, js string) interface{} {
+		prepareFunc: func(t *testing.T, js interface{}) interface{} {
 			doc, _, err := types.JSON.Convert(js)
 			require.NoError(t, err)
 			val, err := doc.(sql.JSONWrapper).ToInterface()
@@ -54,7 +54,7 @@ var jsonFormatTests = []jsonFormatTest{
 	},
 	{
 		name: "LazyJsonDocument",
-		prepareFunc: func(t *testing.T, js string) interface{} {
+		prepareFunc: func(t *testing.T, js interface{}) interface{} {
 			doc, _, err := types.JSON.Convert(js)
 			require.NoError(t, err)
 			bytes, err := types.MarshallJson(doc.(sql.JSONWrapper))
@@ -65,11 +65,11 @@ var jsonFormatTests = []jsonFormatTest{
 }
 
 type testCase struct {
-	name     string
 	f        sql.Expression
 	row      sql.Row
 	expected interface{}
 	err      error
+	name     string
 }
 
 func buildGetFieldExpressions(t *testing.T, construct func(...sql.Expression) (sql.Expression, error), argCount int) sql.Expression {
@@ -82,179 +82,6 @@ func buildGetFieldExpressions(t *testing.T, construct func(...sql.Expression) (s
 	require.NoError(t, err)
 
 	return result
-}
-
-func JsonInsertTestCases(t *testing.T, prepare prepareJsonValue) []testCase {
-
-	jsonInput := prepare(t, `{"a": 1, "b": [2, 3], "c": {"d": "foo"}}`)
-
-	f1 := buildGetFieldExpressions(t, json.NewJSONInsert, 3)
-	f2 := buildGetFieldExpressions(t, json.NewJSONInsert, 5)
-
-	return []testCase{
-		{
-			name:     "insert into beginning of top-level object",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.A", 10.1},
-			expected: `{"A": 10.1, "a": 1, "b": [2, 3], "c": {"d": "foo"}}`,
-		},
-		{
-			name:     "insert at end of top-level object",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.z", 10.1},
-			expected: `{"a": 1, "b": [2, 3], "c": {"d": "foo"}, "z": 10.1}`,
-		},
-		{
-			name:     "insert in middle of top-level object",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.bb", 10.1},
-			expected: `{"a": 1, "b": [2, 3], "bb": 10.1, "c": {"d": "foo"}}`,
-		},
-		{
-			name:     "insert to non-existent path is a no-op",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.bb.cc", 10.1},
-			expected: jsonInput,
-		},
-		{
-			name:     "insert existing does nothing",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.a", 10.1},
-			expected: jsonInput,
-		},
-		{
-			name:     "insert existing nested does nothing",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.c.d", "test"},
-			expected: jsonInput,
-		},
-		{
-			name:     "insert multiple, one change",
-			f:        f2,
-			row:      sql.Row{jsonInput, "$.a", 10.1, "$.e", "new"},
-			expected: `{"a": 1, "b": [2, 3], "c": {"d": "foo"},"e":"new"}`,
-		},
-		{
-			name:     "insert nested does nothing",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.a.e", "test"},
-			expected: jsonInput,
-		},
-		{
-			name:     "insert nested in existing struct",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.c.e", "test"},
-			expected: `{"a": 1, "b": [2, 3], "c": {"d": "foo","e":"test"}}`,
-		},
-		{
-			name:     "insert struct with indexing out of range",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.c[5]", 4.1},
-			expected: `{"a": 1, "b": [2, 3], "c": [{"d": "foo"}, 4.1]}`},
-		{
-			name:     "insert element in array does nothing",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.b[0]", 4.1},
-			expected: jsonInput,
-		},
-		{
-			name:     "insert element in array out of range",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.b[5]", 4.1},
-			expected: `{"a": 1, "b": [2, 3, 4.1], "c": {"d": "foo"}}`,
-		},
-		{
-			name:     "insert nested in array does nothing",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.b.c", 4},
-			expected: jsonInput,
-		},
-		{
-			name:     "struct as array does nothing",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$.a[0]", 4.1},
-			expected: jsonInput,
-		},
-		{
-			name:     "struct does nothing",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$[0]", 4.1},
-			expected: jsonInput,
-		},
-		{
-			name: "improper struct indexing",
-			f:    f1,
-			row:  sql.Row{jsonInput, "$.[0]", 4.1},
-			err:  fmt.Errorf("Invalid JSON path expression. Expected field name after '.' at character 2 of $.[0]"),
-		},
-		{
-			name: "invalid path",
-			f:    f1,
-			row:  sql.Row{jsonInput, "foo", "test"},
-			err:  fmt.Errorf("Invalid JSON path expression. Path must start with '$'"),
-		},
-		{
-			name: "path contains * wildcard",
-			f:    f1,
-			row:  sql.Row{jsonInput, "$.c.*", "test"},
-			err:  fmt.Errorf("Invalid JSON path expression. Expected field name after '.' at character 4 of $.c.*"),
-		},
-		{
-			name: "path contains ** wildcard",
-			f:    f1,
-			row:  sql.Row{jsonInput, "$.c.**", "test"},
-			err:  fmt.Errorf("Invalid JSON path expression. Expected field name after '.' at character 4 of $.c.**"),
-		},
-		{
-			name: "path contains ** wildcard",
-			f:    f1,
-			row:  sql.Row{1, "$.c.**", "test"},
-			err:  sql.ErrInvalidJSONArgument.New(1, "json_insert"),
-		},
-		{
-			name: "path contains ** wildcard",
-			f:    f1,
-			row:  sql.Row{`()`, "$.c.**", "test"},
-			err:  sql.ErrInvalidJSONText.New(1, "json_insert", "()"),
-		},
-		{
-			name:     "whole document no opt",
-			f:        f1,
-			row:      sql.Row{jsonInput, "$", 10.1},
-			expected: jsonInput,
-		},
-		{
-			name:     "sql-null document returns sql-null",
-			f:        f1,
-			row:      sql.Row{nil, "$", 42.7},
-			expected: nil,
-		},
-		{
-			name:     "json-null document returns json-null",
-			f:        f1,
-			row:      sql.Row{"null", "$", 42.7},
-			expected: "null",
-		},
-		{
-			name: "if any path is null, return null",
-			f:    f1,
-			row:  sql.Row{jsonInput, nil, 10},
-		},
-
-		// mysql> select JSON_INSERT(JSON_ARRAY(), "$[2]", 1 , "$[2]", 2 ,"$[2]", 3 ,"$[2]", 4);
-		// +------------------------------------------------------------------------+
-		// | JSON_INSERT(JSON_ARRAY(), "$[2]", 1 , "$[2]", 2 ,"$[2]", 3 ,"$[2]", 4) |
-		// +------------------------------------------------------------------------+
-		// | [1, 2, 3]                                                              |
-		// +------------------------------------------------------------------------+
-		{f: buildGetFieldExpressions(t, json.NewJSONInsert, 9),
-			row: sql.Row{`[]`,
-				"$[2]", 1.1, // [] -> [1.1]
-				"$[2]", 2.2, // [1.1] -> [1.1,2.2]
-				"$[2]", 3.3, // [1.1, 2.2] -> [1.1, 2.2, 3.3]
-				"$[2]", 4.4}, // [1.1, 2.2, 3.3] -> [1.1, 2.2, 3.3]
-			expected: `[1.1, 2.2, 3.3]`},
-	}
 }
 
 func RunJsonTests(t *testing.T, testCases []testCase) {
@@ -290,7 +117,7 @@ func RunJsonTests(t *testing.T, testCases []testCase) {
 				}
 			} else {
 				req.Error(err, "Expected an error but got %v", result)
-				req.Equal(tstC.err.Error(), err.Error())
+				req.Contains(err.Error(), tstC.err.Error())
 			}
 		})
 	}
